lzc_set_prop
Sets a single property of a single dataset.

Atomically setting multiple properties is not implemented in the kernel yet
and that would be quite hard to implement given possible inter-dependencies
between the properties.

Potentially this interface could be extended to work on multiple datasets
at once but it is not clear if that is often needed in practice.

A note on the interface: nvlist_t is used here because there is no API
to create and manage a standalone nvpair_t.  And using explicit value
types would require multiple variants of lzc_set_prop or, at least, passing
the value via void* and its type as a separate variable.