Implement new-style special property handling
Currently, we have two types of properties. General properties that can
be trivially updated in a single DMU transaction and special properties
that implement their own handling and DMU transactions.  Setting
multiple properties can be set simultaneously requires that the special
properties be refactored to share the DMU transaction of the regular
properties.

At present, the properties quota, refquota, reservation and
refreservation are updated in the dsl_sync_task context much like the
regular properties. They can be trivially refactored to be done in the
same transaction group as the others, so we refactor them.

The remaining snapdev, version, volsize and user properties are still
handled separately. Refactoring them is non-trivial in comparison. To
provide some specific points:

volsize requires that calculate how much space needs to be freed when
volume sizes are decreased in advance of calling `dsl_sync_task()`. If
we have concurrent writers, this is non-trivial to do.

version requires that we calculate how much space is needed when system
attributes are added. It is not clear to me what the correct number is,
so I err on the side of caution.

snapdev has a TOCTOU race between the check and sync functions where the
device might be opened. Attempting to set this in a manner intended to
atomically set properties when we cannot atomically guarentee that we
can always remove the device is an inconsistent behavior.

The user properties were not inspected due to time constraints (and how
rarely they are used).

Signed-off-by: Richard Yao <ryao@gentoo.org>