lzc_receive_with_header: allow caller to access information in begin header
lzc_receive_with_header is like lzc_receive, but allows the caller to read
the begin record and then to pass it in.
That could be useful if the caller wants to derive, for example,
the snapname or the origin parameters based on the information contained
in the begin record. The begin record must be in its original form as read
from the stream, in other words, it should not be byteswapped.

'resumable' parameter does nothing and is added for forward compatibility.

lzc_receive_with_header rejects NULL value of begin_record parameter.
This function is specifically designed to be called after the
header has been read and so the header must be passed in.
With NULL begin_record the implementation would try to read
the header again.