zfs: call recv_incremental_replication only for recusrive streams
recv_incremental_replication() takes care of things like removing
datasets that have been removed on the sending side, detecting
renamed datasets, ensuring that all datasets in the affected
hierarchy have the same properties as their counterparts on the
sending side.
All of the above are not necessary if we are receiving a stream
for a single dataset that has been generated with zfs send -p,
that is, a stream that includes properties.
zfs_receive_one() already takes care of applying the properties
to the received datasets.

zfs(1) documentation needs to be updated to spell out how exactly
a stream created with zfs send -p ... gets received.